+++ qexpy/core/__init__.py
@@ -1,0 +1,3 @@
+"""Defines the core data structures of QExPy"""
+
+from .measurement import Measurement