+++ qexpy/core/measurement.py
@@ -1,0 +1,9 @@
+"""Defines measurements"""
+
+from __future__ import annotations
+
+from .experimental_value import ExperimentalValue
+
+
+class Measurement(ExperimentalValue):
+    pass